+++ app/controllers/decidim/decidim_awesome/admin/admin_accountability_controller.rb
@@ -43,8 +43,8 @@ module Decidim
         end
 
         def admin_role_actions
-          @admin_ASrole_actions ||= PaperTrailVersion.in_organization(current_organization)
-                                                     .admin_role_actions(PaperTrailVersion.safe_admin_role_type(params[:admin_role_type]))
+          @admin_role_actions ||= PaperTrailVersion.in_organization(current_organization)
+                                                   .admin_role_actions(PaperTrailVersion.safe_admin_role_type(params[:admin_role_type]))
         end
 
         def export_params
